Splitting models cases and tests
